feat: add navigation and days
## What it does:
This commit adds dynamic rendering for days, months, and years. It also provides navigation among the dates.

## Files and Folders:
`CalendarApp.jsx` uses the `useState` hook to set the current month and year in a way we can further get all the days in the specific month being displayed, as well as set the first day of the month, creating empty `span`s before it, granting it the traditional design we all know. Regarding the transition between months, the code utilizes an `onClick` event, allowing us to set the proper set of days based on the previously explained logic. The current day is also implemented by customizing the `span` that represents it.